chore: restructure to add sub commands to cli
