+++ src/loader.ts
@@ -1,0 +1,162 @@
+/**
+ * MIT License
+ *
+ * Copyright (c) 2020-present, Elastic NV
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ */
+
+import { stdin, cwd } from 'process';
+import { resolve } from 'path';
+import { CliArgs } from './common_types';
+import { step, journey } from './core';
+import { log } from './core/logger';
+import { expect } from './core/expect';
+import {
+  isDepInstalled,
+  isDirectory,
+  totalist,
+  findPkgJsonByTraversing,
+} from './helpers';
+
+const resolvedCwd = cwd();
+
+export async function loadTestFiles(options: CliArgs, args: string[]) {
+  /**
+   * Transform `.ts` files out of the box by invoking
+   * the `ts-node` via `transpile-only` mode that compiles
+   * TS files without doing any extensive type checks.
+   *
+   * We must register `ts-node` _before_ loading inline
+   * scripts too because otherwise we will not be able to
+   * require `.ts` configuration files.
+   */
+  /* eslint-disable-next-line @typescript-eslint/no-var-requires */
+  require('ts-node').register({
+    transpileOnly: true,
+    compilerOptions: {
+      esModuleInterop: true,
+      allowJs: true,
+      target: 'es2020',
+    },
+  });
+
+  /**
+   * Preload modules before running the tests
+   */
+  const modules = [].concat(options.require || []).filter(Boolean);
+  for (const name of modules) {
+    if (isDepInstalled(name)) {
+      require(name);
+    } else {
+      throw new Error(`cannot find module '${name}'`);
+    }
+  }
+
+  if (options.inline) {
+    const source = await readStdin();
+    loadInlineScript(source);
+    return;
+  }
+  /**
+   * Handle piped files by reading the STDIN
+   * ex: ls example/suites/*.js | npx @elastic/synthetics
+   */
+  const files =
+    args.length > 0 ? args : (await readStdin()).split('\n').filter(Boolean);
+  const suites = await prepareSuites(files, options);
+  requireSuites(suites);
+}
+
+const loadInlineScript = source => {
+  const scriptFn = new Function(
+    'step',
+    'page',
+    'context',
+    'browser',
+    'params',
+    'expect',
+    source
+  );
+  journey('inline', ({ page, context, browser, params }) => {
+    scriptFn.apply(null, [step, page, context, browser, params, expect]);
+  });
+};
+
+/**
+ * Read the input from STDIN and run it as inline journey
+ */
+async function readStdin() {
+  const chunks = [];
+  stdin.resume();
+  stdin.setEncoding('utf-8');
+  for await (const chunk of stdin) {
+    chunks.push(chunk);
+  }
+  return chunks.join();
+}
+
+function requireSuites(suites: Iterable<string>) {
+  for (const suite of suites) {
+    require(suite);
+  }
+}
+
+/**
+ * Handle both directory and files that are passed through TTY
+ * and add them to suites
+ */
+async function prepareSuites(inputs: string[], cliArgs: CliArgs) {
+  const suites = new Set<string>();
+  const addSuite = absPath => {
+    log(`Processing file: ${absPath}`);
+    suites.add(require.resolve(absPath));
+  };
+  /**
+   * Match all files inside the directory with the
+   * .journey.{mjs|cjs|js|ts) extensions
+   */
+  const pattern = cliArgs.pattern
+    ? new RegExp(cliArgs.pattern, 'i')
+    : /.+\.journey\.([mc]js|[jt]s?)$/;
+  /**
+   * Ignore node_modules by default when running suites
+   */
+  const ignored = /node_modules/i;
+
+  for (const input of inputs) {
+    const absPath = resolve(resolvedCwd, input);
+    /**
+     * Validate for package.json file before running
+     * the suites
+     */
+    findPkgJsonByTraversing(absPath, resolvedCwd);
+    if (isDirectory(absPath)) {
+      await totalist(absPath, (rel, abs) => {
+        if (pattern.test(rel) && !ignored.test(rel)) {
+          addSuite(abs);
+        }
+      });
+    } else {
+      addSuite(absPath);
+    }
+  }
+  return suites.values();
+}
